+++ python/uatk_spc/reader.py
@@ -196,7 +196,7 @@ class Reader:
                 self.time_use_diaries,
                 pl.int_range(0, self.time_use_diaries.shape[0], eager=True)
                 .rename("index")
-                .cast(pl.UInt64)
+                .cast(pl.UInt32)
                 .to_frame(),
             ],
             how="horizontal",
